requirements.txt
parameterized==0.7.0
boto3==1.9.243
pytz==2019.1
python-digitalocean==1.14.0
